refactor : Period 컴포넌트 삭제, PeriodBox 컴포넌트 분리
